58: migrate DataExplorer component to Vue3
